force it to comply
