+++ pyproject.toml
@@ -100,6 +100,8 @@ asyncio_mode = "auto"
 omit = [
     "**/tests/*",
     "src/migrations/*",
+    "src/db/fields.py",
+    "src/db/models.py",
 ]
 concurrency = ["thread", "greenlet"]
