+++ src/db/models.py
@@ -1,16 +1,14 @@
 import uuid
 from datetime import datetime
-from typing import Any, ClassVar, Literal, TypeVar, override
+from typing import Any, ClassVar, override
 
-from sqlalchemy import JSON, DateTime, ForeignKey, MetaData, func
-from sqlalchemy.dialects.postgresql import UUID
+from sqlalchemy import JSON, DateTime, MetaData, func
 from sqlalchemy.ext.asyncio import AsyncAttrs
 from sqlalchemy.orm import (
     DeclarativeBase,
     Mapped,
     declared_attr,
     mapped_column,
-    relationship,
 )
 from sqlalchemy.types import TypeEngine
 
@@ -39,27 +37,3 @@ class Base(AsyncAttrs, DeclarativeBase):
     @override
     def __tablename__(cls) -> str:
         return f"{cls.__name__.lower()}s"
-
-
-Model = TypeVar("Model", bound=Base)
-
-
-def foreign_key(target: type[Base], ondelete: Literal["CASCADE", "SET NULL"] = "CASCADE") -> Mapped[uuid.UUID]:
-    return mapped_column(UUID(as_uuid=True), ForeignKey(target.id, ondelete=ondelete))
-
-
-def parent_relationship(target: type[Model], *, back_populates: str) -> Mapped[Model]:
-    return relationship(target, back_populates=back_populates)
-
-
-def child_relationship(
-    target: str, *, back_populates: str, uselist: bool | None = None, order_by: str | None = None
-) -> Mapped[list[Model]]:
-    return relationship(
-        target,
-        back_populates=back_populates,
-        uselist=uselist,
-        cascade="all, delete-orphan",
-        passive_deletes=True,
-        order_by=order_by or f"{target}.id",
-    )
